Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/cssom-getBoundingClientRect-003-expected.txt

XXX

FAIL {Element,Range}.prototype.getBoundingClientRect on SVG <tspan>, Element assert_equals: left expected 108 but got 8
FAIL {Element,Range}.prototype.getBoundingClientRect on SVG <tspan>, Range assert_equals: top expected 108 but got 107.984375
